Add fixed buffers support

+++ src/runtime/config.rs
@@ -31,13 +31,13 @@ impl ConfigStats {
 /// A shared config of state of the all runtime.
 static GLOBAL_CONFIG_STATS: SpinLock<ConfigStats> = SpinLock::new(ConfigStats::new());
 
-/// The default [`buffers`](crate::buf::Buffer) capacity.
-pub const DEFAULT_BUF_CAP: usize = 4096;
+/// The default [`buffers`](crate::io::Buffer) capacity.
+pub const DEFAULT_BUF_CAP: u32 = 4096;
 
 /// Config that can be used to create an Executor, because it is valid.
 #[derive(Clone)]
 pub(crate) struct ValidConfig {
-    pub(crate) buffer_cap: usize,
+    pub(crate) buffer_cap: u32,
     pub(crate) io_worker_config: Option<IoWorkerConfig>,
     pub(crate) number_of_thread_workers: usize,
     /// If it is `usize::MAX`, it means that work sharing is disabled.
@@ -84,7 +84,7 @@ impl Drop for ValidConfig {
 /// related to buffers, I/O workers, thread workers, and work-sharing behavior.
 ///
 /// # Fields
-/// - `buffer_cap`: The size of the [`buffers`](crate::buf::Buffer).
+/// - `buffer_cap`: The size of the [`buffers`](crate::io::Buffer).
 ///
 /// - `io_worker_config`: An optional configuration for I/O workers. If none is provided,
 ///   the IO worker will be disabled.
@@ -98,8 +98,8 @@ impl Drop for ValidConfig {
 ///   If [`usize::MAX`] is provided, work sharing will be disabled.
 #[derive(Clone, Copy)]
 pub struct Config {
-    /// The size of the [`buffers`](crate::buf::Buffer).
-    buffer_cap: usize,
+    /// The size of the [`buffers`](crate::io::Buffer).
+    buffer_cap: u32,
     /// An optional configuration for I/O workers. If none is provided,
     /// the IO worker will be disabled.
     io_worker_config: Option<IoWorkerConfig>,
@@ -152,14 +152,14 @@ impl Config {
         }
     }
 
-    /// Returns the capacity of the [`buffers`](crate::buf::Buffer).
-    pub const fn buffer_cap(&self) -> usize {
+    /// Returns the capacity of the [`buffers`](crate::io::Buffer).
+    pub const fn buffer_cap(&self) -> u32 {
         self.buffer_cap
     }
 
-    /// Sets the capacity of the [`buffers`](crate::buf::Buffer).
+    /// Sets the capacity of the [`buffers`](crate::io::Buffer).
     #[must_use]
-    pub const fn set_buffer_cap(mut self, buf_cap: usize) -> Self {
+    pub const fn set_buffer_cap(mut self, buf_cap: u32) -> Self {
         self.buffer_cap = buf_cap;
 
         self
